all_tests_clean

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: marferre <marferre@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/12 15:36:04 by marferre          #+#    #+#             */
-/*   Updated: 2022/09/12 18:20:01 by marferre         ###   ########.fr       */
+/*   Updated: 2022/09/25 19:27:32 by marferre         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
